backend/requirements.txt
fastapi
uvicorn
python-multipart
PyJWT
openai
# django (kept for reference or if needed for some utilities, but eventually can be removed)
django
djangorestframework
django-cors-headers
